# openapi.yaml
openapi: 3.0.3
info:
  title: Push Balancer API
  description: |
    ML-gestützte API für BILD Push-Notification-Optimierung.

    Der Push Balancer analysiert historische Push-Daten, prognostiziert Opening Rates (OR)
    via Gradient Boosted Trees und Stacking Ensemble (LightGBM + XGBoost + CatBoost → Ridge),
    und liefert tagesaktuelle Empfehlungen für Redaktion und Sport.

    **Safety**: Alle Vorhersagen sind ausschließlich beratend (ADVISORY_ONLY). Das System
    darf niemals autonom Push-Benachrichtigungen versenden. Jede Prediction enthält
    `advisoryOnly: true` und `actionAllowed: false`.
  version: 2.4.0
  contact:
    name: BILD Digital Produkt
    email: push-balancer@axelspringer.com
servers:
  - url: http://localhost:8050
    description: Lokaler Entwicklungsserver
  - url: https://push-balancer.onrender.com
    description: Render Cloud Deployment
...
paths:
...
  /api/push/{proxyPath}:
    get:
      operationId: getPushApiProxy
      tags: [push]
      summary: Push-API Proxy
      description: |
        Transparenter Proxy zur internen BILD Push-Statistics-API
        (`push-frontend.bildcms.de`). Der Pfad nach `/api/push/` wird direkt
        an die Upstream-API weitergeleitet. Bei Nichterreichbarkeit wird auf
        den lokalen Sync-Cache zurückgegriffen.
      parameters:
        - name: proxyPath
          in: path
          required: true
          description: Pfad, der an die Push-Statistics-API weitergeleitet wird
          schema:
            type: string
            example: "pushes/today"
      responses:
        "200":
          description: Antwort der Push-Statistics-API (JSON)
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "502":
          $ref: "#/components/responses/BadGateway"
  /api/push-sync:
    post:
      operationId: syncPushData
      tags: [push]
      summary: Push-Daten synchronisieren
      description: |
        Empfängt Push-Daten vom lokalen Produktivserver und speichert sie im
        In-Memory-Cache (für Render-Deployments ohne direkten Zugang zur
        internen Push-API). Erfordert den konfigurierten `secret`-Schlüssel.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PushSyncRequest"
      responses:
        "200":
          description: Sync erfolgreich
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                    example: true
                  received:
                    type: integer
                    description: Anzahl empfangener Nachrichten
                    example: 47
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalServerError"

  # ── ML ───────────────────────────────────────────────────────────────────
...
  /api/prediction-feedback:
    post:
      operationId: submitPredictionFeedback
      tags: [ml]
      summary: Prediction-Feedback einreichen
      description: |
        Meldet das tatsächliche Ergebnis eines Pushes zurück (actual OR vs. predicted OR).
        Wird für Online-Bias-Korrektur und A/B-Test-Auswertung genutzt. Duplikate
        (gleiche `pushId`) werden ignoriert.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PredictionFeedbackRequest"
      responses:
        "200":
          description: Feedback gespeichert
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    enum: [ok, duplicate]
                    example: ok
                  count:
                    type: integer
                    description: Gesamtanzahl gespeicherter Feedback-Einträge
                    example: 142
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalServerError"

  # ── GBRT ─────────────────────────────────────────────────────────────────
...
  /api/tagesplan:
    get:
      operationId: getTagesplan
      tags: [tagesplan]
      summary: Heutigen Push-Tagesplan abrufen
      description: |
        Liefert den ML-generierten Tagesplan für heute: 18 Stunden-Slots (06:00–23:00)
        mit empfohlenen Kategorien, prognostizierten ORs, bereits gesendeten Pushes
        und Top-Artikelvorschlägen. Unterstützt Modi `redaktion` und `sport`.
      parameters:
        - name: mode
          in: query
          required: false
          description: "Planungsmodus: `redaktion` für allgemeine Pushes, `sport` für Sport-Fokus"
          schema:
            type: string
            enum: [redaktion, sport]
            default: redaktion
      responses:
        "200":
          description: Heutiger Tagesplan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TagesplanResponse"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/tagesplan/retro:
    get:
      operationId: getTagesplanRetro
      tags: [tagesplan]
      summary: 7-Tage-Retrospektive abrufen
      description: |
        Liefert eine aggregierte Retrospektive der letzten 7 Tage: durchschnittliche
        OR pro Tag, beste und schlechteste Stunden, häufigste Kategorien.
      responses:
        "200":
          description: 7-Tage-Retrospektive
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TagesplanRetroResponse"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/tagesplan/history:
    get:
      operationId: getTagesplanHistory
      tags: [tagesplan]
      summary: Tagesplan für vergangenes Datum abrufen
      description: |
        Liefert den vollständigen Tagesplan für ein vergangenes Datum aus der
        SQLite-Datenbank: 18 Stunden-Slots mit tatsächlich gesendeten Pushes,
        deren echten ORs, prognostizierten ORs (Prediction-Log) und den damaligen
        Artikelvorschlägen.
      parameters:
        - name: date
          in: query
          required: true
          description: Datum im Format YYYY-MM-DD
          schema:
            type: string
            format: date
            example: "2026-03-28"
        - name: mode
          in: query
          required: false
          description: "Planungsmodus: `redaktion` oder `sport`"
          schema:
            type: string
            enum: [redaktion, sport]
            default: redaktion
      responses:
        "200":
          description: Historischer Tagesplan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TagesplanHistoryResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/tagesplan/suggestions:
    get:
      operationId: getTagesplanSuggestions
      tags: [tagesplan]
      summary: Gespeicherte Artikelvorschläge abrufen
      description: |
        Liefert die vom Frontend geloggten Artikelvorschläge pro Stunden-Slot für
        ein bestimmtes Datum. Gibt ein Objekt zurück, bei dem die Slot-Stunde (als
        String) auf eine Liste von Vorschlägen gemappt ist.
      parameters:
        - name: date
          in: query
          required: false
          description: "Datum im Format YYYY-MM-DD (Standard: heute)"
          schema:
            type: string
            format: date
            example: "2026-04-02"
        - name: mode
          in: query
          required: false
          description: Planungsmodus (wird aktuell nicht gefiltert, reserviert)
          schema:
            type: string
            enum: [redaktion, sport]
            default: redaktion
      responses:
        "200":
          description: Artikelvorschläge nach Slot-Stunde
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TagesplanSuggestionsResponse"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/tagesplan/log-suggestions:
    post:
      operationId: logTagesplanSuggestions
      tags: [tagesplan]
      summary: Artikelvorschläge loggen
      description: |
        Das Frontend loggt die aktuell angezeigten Artikelvorschläge pro Tagesplan-Slot
        für spätere Gegenuberstellung mit den tatsächlich versendeten Pushes.
        Max. 36 Einträge (18 Slots × 2 Vorschläge). Einträge älter als 30 Tage
        werden automatisch gelöscht.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LogSuggestionsRequest"
      responses:
        "200":
          description: Vorschläge geloggt
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
                  logged:
                    type: integer
                    description: Anzahl gespeicherter Vorschläge
                    example: 18
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalServerError"

  # ── AI / Chat ─────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Problem:
      type: object
      description: RFC 9457 Problem Details für HTTP APIs
      required:
        - type
        - title
        - status
      properties:
        type:
          type: string
          format: uri
          description: URI-Referenz, die den Fehlertyp identifiziert
          example: "https://push-balancer.example.com/problems/validation-error"
        title:
          type: string
          description: Kurze, menschenlesbare Fehlerbeschreibung
          example: "Validation Error"
        status:
          type: integer
          description: HTTP-Statuscode
          example: 400
        detail:
          type: string
          description: Ausführlichere Fehlerbeschreibung
          example: "Das Feld 'pushId' darf nicht leer sein"
        instance:
          type: string
          format: uri
          description: URI-Referenz des konkreten Fehler-Vorkommens
          example: "/api/prediction-feedback"

    # ── Health ──────────────────────────────────────────────────────────────
...
    MlMetrics:
      type: object
      properties:
        mae:
          type: number
          format: float
          description: Mean Absolute Error in Prozentpunkten
          example: 1.42
        rmse:
          type: number
          format: float
          description: Root Mean Squared Error in Prozentpunkten
          example: 2.01
        r2:
          type: number
          format: float
          description: R²-Score (Bestimmtheitsmaß)
          example: 0.71
        trainN:
          type: integer
          description: Anzahl Trainings-Samples
          example: 2800
        valN:
          type: integer
          description: Anzahl Validierungs-Samples
          example: 600
        testN:
          type: integer
          description: Anzahl Test-Samples
          example: 400
...
    PredictionFeedbackRequest:
      type: object
      required: [pushId, predictedOr, actualOr]
      properties:
        pushId:
          type: string
          description: Eindeutige Push-ID
          example: "msg_12345"
        predictedOr:
          type: number
          format: float
          description: Vom Modell prognostizierte Opening Rate
          example: 7.24
        actualOr:
          type: number
          format: float
          description: Tatsächlich gemessene Opening Rate (nach Reife >24h)
          example: 6.87
        basisMethod:
          type: string
          description: Verwendete Methode (z.B. gbrt, ensemble, fallback)
          example: ensemble
        pushTitle:
          type: string
          description: Push-Titelzeile
          example: "So reagiert Deutschland auf die Wahl"
        pushCat:
          type: string
          description: Artikel-Kategorie
          example: Politik
        pushHour:
          type: integer
          description: Versand-Stunde
          example: 18
        source:
          type: string
          description: Herkunft des Feedbacks
          example: frontend
        championPred:
          type: number
          format: float
          nullable: true
          description: Champion-Modell-Prognose (für A/B-Test)
        challengerPred:
          type: number
          format: float
          nullable: true
          description: Challenger-Modell-Prognose (für A/B-Test)

    # ── GBRT ────────────────────────────────────────────────────────────────
...
    TagesplanPush:
      type: object
      description: Tatsächlich versendeter Push in einem Stunden-Slot
      properties:
        title:
          type: string
          description: Push-Titelzeile
          example: "Bayern siegt im Finale"
        cat:
          type: string
          description: Kategorie
          example: sport
        or:
          type: number
          format: float
          description: Tatsächliche Opening Rate in Prozent (aus DB)
          example: 8.42
        predictedOr:
          type: number
          format: float
          nullable: true
          description: Zum Versandzeitpunkt prognostizierte OR
          example: 7.91
        delta:
          type: number
          format: float
          nullable: true
          description: Differenz predicted − actual OR
          example: -0.51
        isEilmeldung:
          type: boolean
          example: false
        link:
          type: string
          format: uri
          example: "https://www.bild.de/sport/..."
        hour:
          type: integer
          example: 20
    TagesplanSuggestion:
      type: object
      description: Empfohlener Artikel für einen Tagesplan-Slot
      properties:
        num:
          type: integer
          description: Vorschlags-Nummer (1 = bester)
          example: 1
        title:
          type: string
          example: "Überraschung im Endspiel: Bayern besiegt Dortmund"
        link:
          type: string
          format: uri
        category:
          type: string
          example: Sport
        score:
          type: number
          format: float
          description: Push-Score (Eignungsbewertung, 0–100)
          example: 87.3
        expectedOr:
          type: number
          format: float
          description: Prognostizierte Opening Rate für diesen Artikel
          example: 7.84
        bestCat:
          type: string
          description: Empfohlene Best-Kategorie für diesen Slot
          example: Sport
    TagesplanSlot:
      type: object
      description: Einzelner Stunden-Slot im Tagesplan
      properties:
        hour:
          type: integer
          description: Stunde (6–23)
          example: 18
        isPast:
          type: boolean
          description: Liegt dieser Slot in der Vergangenheit?
          example: false
        isNow:
          type: boolean
          description: Ist dies der aktuelle Stunden-Slot?
          example: true
        expectedOr:
          type: number
          format: float
          description: Prognostizierte OR für diesen Slot
          example: 7.2
        bestCat:
          type: string
          description: Empfohlene Kategorie für diesen Slot
          example: Politik
        nPushed:
          type: integer
          description: Anzahl bereits versendeter Pushes in diesem Slot
          example: 1
        pushedThisHour:
          type: array
          description: Tatsächlich versendete Pushes in diesem Slot
          items:
            $ref: "#/components/schemas/TagesplanPush"
    TagesplanResponse:
      type: object
      properties:
        date:
          type: string
          description: Datum im Format DD.MM.YYYY
          example: "02.04.2026"
        dateIso:
          type: string
          format: date
          example: "2026-04-02"
        weekday:
          type: string
          description: Wochentag-Kürzel (Mo–So)
          example: "Do"
        currentHour:
          type: integer
          description: Aktuelle Stunde (0–23)
          example: 18
        slots:
          type: array
          description: 18 Stunden-Slots (06:00–23:00)
          items:
            $ref: "#/components/schemas/TagesplanSlot"
        alreadyPushedToday:
          type: array
          description: Liste aller heute bereits versendeten Pushes
          items:
            $ref: "#/components/schemas/TagesplanPush"
        nPushedToday:
          type: integer
          example: 4
        avgOrToday:
          type: number
          format: float
          description: Durchschnittliche OR aller heutigen Pushes
          example: 6.93
        bestPush:
          $ref: "#/components/schemas/TagesplanPush"
          nullable: true
        goldenHour:
          type: integer
          nullable: true
          description: Empfohlene Top-Stunde heute
          example: 18
        mustHaveHours:
          type: array
          description: Stunden mit dringender Push-Empfehlung
          items:
            type: integer
        totalPushesDb:
          type: integer
          description: Gesamtanzahl Pushes in der Datenbank
          example: 3820
        mlTrained:
          type: boolean
          description: Ob das ML-Modell verfügbar ist
          example: true
        mlMetrics:
          $ref: "#/components/schemas/MlMetrics"
    TagesplanHistoryResponse:
      type: object
      description: Historischer Tagesplan für ein vergangenes Datum
      allOf:
        - $ref: "#/components/schemas/TagesplanResponse"
        - type: object
          properties:
            isHistory:
              type: boolean
              example: true
            suggestions:
              type: object
              description: Map von Slot-Stunde (String) zu Vorschlägen
              additionalProperties:
                type: array
                items:
                  $ref: "#/components/schemas/TagesplanSuggestion"
    TagesplanRetroResponse:
      type: object
      properties:
        days:
          type: array
          description: Tages-Zusammenfassungen der letzten 7 Tage
          items:
            type: object
            properties:
              dateIso:
                type: string
                format: date
              date:
                type: string
                example: "Di, 01.04.2026"
              avgOr:
                type: number
                format: float
              nPushed:
                type: integer
              bestHour:
                type: integer
                nullable: true
              bestCat:
                type: string
    TagesplanSuggestionsResponse:
      type: object
      properties:
        dateIso:
          type: string
          format: date
          example: "2026-04-02"
        suggestions:
          type: object
          description: Map von Slot-Stunde (String) zu Vorschlägen
          additionalProperties:
            type: array
            items:
              $ref: "#/components/schemas/TagesplanSuggestion"
    LogSuggestionsRequest:
      type: object
      required: [dateIso, suggestions]
      properties:
        dateIso:
          type: string
          format: date
          description: Datum im Format YYYY-MM-DD
          example: "2026-04-02"
        suggestions:
          type: array
          description: Liste der Vorschläge (max. 36)
          maxItems: 36
          items:
            type: object
            required: [slotHour, suggestionNum]
            properties:
              slotHour:
                type: integer
                description: Stunden-Slot (6–23)
                example: 18
              suggestionNum:
                type: integer
                description: Vorschlags-Nummer (1 oder 2)
                example: 1
              title:
                type: string
                description: Artikeltitel (max. 200 Zeichen)
              link:
                type: string
                format: uri
              category:
                type: string
                example: Politik
              score:
                type: number
                format: float
                example: 87.3
              expectedOr:
                type: number
                format: float
                example: 7.2
              bestCat:
                type: string
                example: Politik

    # ── Push Sync ───────────────────────────────────────────────────────────
    PushSyncRequest:
      type: object
      required: [secret, messages]
      properties:
        secret:
          type: string
          description: Sync-Schlüssel (aus Umgebungsvariable PUSH_SYNC_SECRET)
          example: "bild-push-sync-2026"
        messages:
          type: array
          description: Liste geparster Push-Nachrichten
          items:
            type: object
            additionalProperties: true
        channels:
          type: array
          description: Liste aktiver Push-Kanäle
          items:
            type: string

    # ── AI / Chat ────────────────────────────────────────────────────────────
...
  responses:
    BadRequest:
      description: Ungültige Anfrage (fehlende oder falsche Parameter)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Problem"
          example:
            type: "https://push-balancer.example.com/problems/bad-request"
            title: "Bad Request"
            status: 400
            detail: "Das Feld 'pushId' darf nicht leer sein"
    Forbidden:
      description: Zugriff verweigert (ungültiger Sync-Secret)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Problem"
          example:
            type: "https://push-balancer.example.com/problems/forbidden"
            title: "Forbidden"
            status: 403
            detail: "Invalid sync secret"
...
    InternalServerError:
      description: Interner Serverfehler
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Problem"
          example:
            type: "https://push-balancer.example.com/problems/internal-error"
            title: "Internal Server Error"
            status: 500
            detail: "ML status error: model not initialized"
    BadGateway:
      description: Upstream-Fehler (externe API nicht erreichbar)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Problem"
          example:
            type: "https://push-balancer.example.com/problems/bad-gateway"
            title: "Bad Gateway"
            status: 502
            detail: "Proxy error: Connection timeout"